+++ src/main/java/com/metalancer/backend/orders/domain/DaySalesReport.java
@@ -1,5 +1,6 @@
 package com.metalancer.backend.orders.domain;
 
+import java.math.BigDecimal;
 import lombok.Builder;
 import lombok.Getter;
 
@@ -8,14 +9,14 @@ public class DaySalesReport {
 
     private final String day;
     private final int totalPrice;
-    private final Double totalPriceUSD;
+    private final BigDecimal totalPriceUSD;
     private final int salesCnt;
 
     @Builder
-    public DaySalesReport(String day, Integer totalPrice, Double totalPriceUSD, int salesCnt) {
+    public DaySalesReport(String day, Integer totalPrice, BigDecimal totalPriceUSD, int salesCnt) {
         this.day = day;
         this.totalPrice = totalPrice != null ? totalPrice : 0;
-        this.totalPriceUSD = totalPriceUSD != null ? totalPriceUSD : 0;
+        this.totalPriceUSD = totalPriceUSD != null ? totalPriceUSD : BigDecimal.valueOf(0);
         this.salesCnt = salesCnt;
     }
 }
