+++ src/main/java/com/metalancer/backend/users/domain/PayedAssets.java
@@ -3,6 +3,7 @@ package com.metalancer.backend.users.domain;
 import com.metalancer.backend.common.constants.OrderStatus;
 import com.metalancer.backend.common.utils.Time;
 import com.metalancer.backend.products.domain.RequestOption;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import lombok.Builder;
 import lombok.Getter;
@@ -17,7 +18,7 @@ public class PayedAssets {
     private String orderProductNo;
     private Long productsId;
     private String title;
-    private Double price;
+    private BigDecimal price;
     private String priceUnit;
     private String sellerNickname;
     private String sellerName;
@@ -32,7 +33,7 @@ public class PayedAssets {
     @Builder
     public PayedAssets(Long payedAssetsId, String orderNo, String orderProductNo, Long productsId,
         String title,
-        Double price, String sellerNickname, String sellerName, String sellerPhone,
+        BigDecimal price, String sellerNickname, String sellerName, String sellerPhone,
         String thumbnail, LocalDateTime purchasedAt, Integer downloadedCnt, String downloadLink,
         OrderStatus orderStatus, String priceUnit) {
         this.payedAssetsId = payedAssetsId;
